feat(cli): whole bunch of stuff
Refactoring of Termivore CLI
Added some fns to utils
Exported types from package
Updated .option args in command
Added generated code for Termivore CLI
Updated documentation
Probably more things I can't remember
